Create Day8_StockBuyAndSell.java
Day 8 :- Stock Buy and Sell to Maximize Profit
Problem Statement : 
Given an array of stock prices where the i-th element represents the price of a stock on day i, the task is to calculate the maximum profit achievable by completing at most one buy-and-sell transaction. If no profit can be made, the output should be 0.