Day27_MergeWithoutExtraSpace.java
Day 27 :- Merge Without Extra Space
Problem Statement :- 
Given two sorted arrays a[] and b[] of size n and m respectively, the task is to merge them in sorted order without using any extra space. Modify a[] so that it contains the first n elements and modify b[] so that it contains the last m elements.